server/requirements.txt
flask-sqlalchemy
psycopg2
flask-admin
flask-cors